+++ cogwheels/helpers/tests/test_helper_attribute_shortcuts.py
@@ -1,7 +1,31 @@
 from unittest.mock import patch
 
+from cogwheels import UnknownSettingNameError
 from cogwheels.helpers import BaseAppSettingsHelper
 from cogwheels.tests.base import AppSettingTestCase
+
+
+class TestDirectAttributeShortcut(AppSettingTestCase):
+
+    @patch.object(BaseAppSettingsHelper, 'get')
+    def test_raises_unknownsettingnameerror_if_no_default_defined(self, mocked_method):
+        expected_message = "'UNKNOWN_SETTING' is not a valid setting name"
+        with self.assertRaisesRegex(UnknownSettingNameError, expected_message):
+            self.appsettingshelper.UNKNOWN_SETTING
+        mocked_method.assert_not_called()
+
+    @patch.object(BaseAppSettingsHelper, 'get')
+    def test_raises_simple_attributeerror_if_non_uppercase_attribute_not_found(self, mocked_method):
+        for attribute_name in (
+            'lower_case_attr',
+            'MIXED_case_attr'
+            'CamelCaseAttr'
+        ):
+            try:
+                getattr(self.appsettingshelper, attribute_name)
+            except AttributeError as e:
+                self.assertNotIsInstance(e, UnknownSettingNameError)
+        mocked_method.assert_not_called()
 
 
 class TestModelsShortcut(AppSettingTestCase):
@@ -16,10 +40,23 @@ class TestModelsShortcut(AppSettingTestCase):
     know there is no such setting).
     """
     @patch.object(BaseAppSettingsHelper, 'get_model')
-    def test_raises_attributeerror_if_no_default_defined(self, mocked_method):
-        expected_message = "'I_DONT_THINK_SO' is not a valid setting name"
-        with self.assertRaisesRegex(AttributeError, expected_message):
-            self.appsettingshelper.models.I_DONT_THINK_SO
+    def test_raises_unknownsettingnameerror_if_no_default_defined(self, mocked_method):
+        expected_message = "'UNKNOWN_SETTING' is not a valid setting name"
+        with self.assertRaisesRegex(UnknownSettingNameError, expected_message):
+            self.appsettingshelper.models.UNKNOWN_SETTING
+        mocked_method.assert_not_called()
+
+    @patch.object(BaseAppSettingsHelper, 'get_model')
+    def test_raises_simple_attributeerror_if_non_uppercase_attribute_not_found(self, mocked_method):
+        for attribute_name in (
+            'lower_case_attr',
+            'MIXED_case_attr'
+            'CamelCaseAttr'
+        ):
+            try:
+                getattr(self.appsettingshelper.models, attribute_name)
+            except AttributeError as e:
+                self.assertNotIsInstance(e, UnknownSettingNameError)
         mocked_method.assert_not_called()
 
     @patch.object(BaseAppSettingsHelper, 'get_model')
@@ -52,10 +89,23 @@ class TestModulesShortcut(AppSettingTestCase):
     know there is no such setting).
     """
     @patch.object(BaseAppSettingsHelper, 'get_module')
-    def test_raises_attributeerror_if_no_default_defined(self, mocked_method):
-        expected_message = "'I_DONT_THINK_SO' is not a valid setting name"
-        with self.assertRaisesRegex(AttributeError, expected_message):
-            self.appsettingshelper.modules.I_DONT_THINK_SO
+    def test_raises_unknownsettingnameerror_if_no_default_defined(self, mocked_method):
+        expected_message = "'UNKNOWN_SETTING' is not a valid setting name"
+        with self.assertRaisesRegex(UnknownSettingNameError, expected_message):
+            self.appsettingshelper.modules.UNKNOWN_SETTING
+        mocked_method.assert_not_called()
+
+    @patch.object(BaseAppSettingsHelper, 'get_module')
+    def test_raises_simple_attributeerror_if_non_uppercase_attribute_not_found(self, mocked_method):
+        for attribute_name in (
+            'lower_case_attr',
+            'MIXED_case_attr'
+            'CamelCaseAttr'
+        ):
+            try:
+                getattr(self.appsettingshelper.modules, attribute_name)
+            except AttributeError as e:
+                self.assertNotIsInstance(e, UnknownSettingNameError)
         mocked_method.assert_not_called()
 
     @patch.object(BaseAppSettingsHelper, 'get_module')
@@ -87,10 +137,23 @@ class TestObjectsShortcut(AppSettingTestCase):
     default value defined
     """
     @patch.object(BaseAppSettingsHelper, 'get_object')
-    def test_raises_attributeerror_if_no_default_defined(self, mocked_method):
-        expected_message = "'I_DONT_THINK_SO' is not a valid setting name"
-        with self.assertRaisesRegex(AttributeError, expected_message):
-            self.appsettingshelper.objects.I_DONT_THINK_SO
+    def test_raises_unknownsettingnameerror_if_setting_not_in_defaults(self, mocked_method):
+        expected_message = "'UNKNOWN_SETTING' is not a valid setting name"
+        with self.assertRaisesRegex(UnknownSettingNameError, expected_message):
+            self.appsettingshelper.objects.UNKNOWN_SETTING
+        mocked_method.assert_not_called()
+
+    @patch.object(BaseAppSettingsHelper, 'get_object')
+    def test_raises_simple_attributeerror_if_non_uppercase_attribute_not_found(self, mocked_method):
+        for attribute_name in (
+            'lower_case_attr',
+            'MIXED_case_attr'
+            'CamelCaseAttr'
+        ):
+            try:
+                getattr(self.appsettingshelper.objects, attribute_name)
+            except AttributeError as e:
+                self.assertNotIsInstance(e, UnknownSettingNameError)
         mocked_method.assert_not_called()
 
     @patch.object(BaseAppSettingsHelper, 'get_object')